Scrapt/Tools/requirements.txt
youtube-comment-downloader
reportlab
matplotlib
tqdm
pandas
colorama
wordcloud
textblob
openpyxl
google-genai